Confirm on Delete

+++ components/Cards/UserTable.js
@@ -33,7 +33,8 @@ export default function UserTable({ user }) {
   }, [user, responseUser]);
 
   const deleteUser = (e, id) => {
-    {alert("Are you sure);}
+    let confirmed = confirm("Are you sure you wanna delete this user?");
+    if (!confirmed) return;
     e.preventDefault();
     fetch(USER_API_BASE_URL + "/" + id, {
       method: "DELETE",
